CPL scripts : several modification for easier use, improved logs and error tracking (fix issue # 153 and other improvements following feedback from CROCO schools)
* TOY : some corrections and simplifications for 2 toys:
    * now one toy executable per model type (oce, atm, wav)
    * a script for compilation is given
    * namcouples have been adapted accordingly

* step PUT : remove ncrcat (too long, useless) and renaming for WRF and WW3 (original names are ok). Files are simply moved to the output dir

* ATM:
    * getfile : simplify by putting all options into one loop better using filelist
    * add check of useful data for wrf
    * getrst : change input files format as wrfinput_d01_YYY_MM_DD_HH, to allows DAY selection (HOUR sill would require additional changes in the scripts)
    * correction for CPLMASK creation when nests
    * real: simpify configure.namelist.real
    * atm_nam : add possibility to add external text files for iofields (but not mandatory)

* change MY_NODES in case of MPI_NOLAND to NP_OCE and put it into myjob instead of mynamelist

* OCE compil :
    * always done when asked with ONLINE_COMP except if restart_flag=TRUE (chaining)
    * compilatio is now done in jobs_CEXPER so that several experiments with different options/compilations may be runned at the same time

* create_oasis_restart : some corrections, and dedicated log

* deal better with jobs launch and myjob automatic modification:
    * every useful script is copied to the jobs_CEXPER directory (myenv, mynamelist, myjob, submitjob) so that the experiment can be re-launched from there.
    * myjob.sh copied as myjob.sh.bck before updating it with new dates
    * next job is launched from jobs_CEXPER to avoid interference with other experiments
    * add the possibility to answer « yes » to remove previous job existing when using submitjob

* mynamelist.sh:
    * cleaning, re-arrangement
    * simplify by better using RUNtype
    * checks on consistency between RUNtype and model options

* improve cosmetics, logs, prints, checks and exits

* job output files *.eXXXX, *.oXXXX move was commented as not relevant as the job is on one step only at the moment. can be re-activated when mutli-steps will be possible

+++ SCRIPTS/SCRIPTS_COUPLING/SCRIPTS_TOOLBOX/wav_nam.sh
@@ -34,18 +34,23 @@ sed -e "s/<yr1>/${YEAR_BEGIN_JOB}/g"  -e "s/<mo1>/${ms}/g"  -e "s/<dy1>/${ds}/g"
     -e "/$     <point_output_list>/r tmp_point_output_list"  \
     -e "s/$    <T>/$my_track_flag/g" \
     -e "s/<wavdt>/${DT_WAV}/g" \
-    ${WAV_NAM_DIR}/ww3_shel.inp.base.${RUNtype} > ./ww3_shel.inp
+    ${WAV_NAM_DIR}/ww3_shel.inp.base.${SHELL_EXT} > ./ww3_shel.inp
 
 echo 'Fill ww3_ounf.inp file'
 sed -e "s/<wav_int>/${wav_int}/g" \
     -e "s/<yr1>/${YEAR_BEGIN_JOB}/g"  -e "s/<mo1>/${ms}/g"  -e "s/<dy1>/${ds}/g"  -e "s/<hr1>/00/g" \
     ${WAV_NAM_DIR}/ww3_ounf.inp.base > ./ww3_ounf.inp
 
-if [ ${wav_pnt} -ne 0 ]; then 
-    echo 'Fill ww3_ounp.inp file'
-    sed -e "s/<wav_pnt>/${wav_pnt}/g" \
-        -e "s/<yr1>/${YEAR_BEGIN_JOB}/g"  -e "s/<mo1>/${ms}/g"  -e "s/<dy1>/${ds}/g"  -e "s/<hr1>/00/g" \
-        ${WAV_NAM_DIR}/ww3_ounp.inp.base > ./ww3_ounp.inp
+if [ ${wav_pnt} -ne 0 ]; then
+    if [ -f ${point_output_list} ] ; then  
+        echo 'Fill ww3_ounp.inp file'
+        sed -e "s/<wav_pnt>/${wav_pnt}/g" \
+            -e "s/<yr1>/${YEAR_BEGIN_JOB}/g"  -e "s/<mo1>/${ms}/g"  -e "s/<dy1>/${ds}/g"  -e "s/<hr1>/00/g" \
+            ${WAV_NAM_DIR}/ww3_ounp.inp.base > ./ww3_ounp.inp
+    else
+         echo "ERROR: wav_pnt is not 0 and point_output_list=${point_output_list} file cannot be found. Exit"
+         exit 1
+    fi
 fi
 
 if [ ! -z $bouncin ]; then
